+++ addons/eu/en16931/bill_test.go
@@ -252,6 +252,15 @@ func TestValidateBillDiscount(t *testing.T) {
 		l := &bill.Discount{
 			Reason: "Product sample",
 			Amount: num.MakeAmount(100, 2),
+			Taxes: tax.Set{
+				{
+					Category: tax.CategoryVAT,
+					Rate:     "standard",
+					Ext: tax.Extensions{
+						untdid.ExtKeyTaxCategory: en16931.TaxCategoryStandard,
+					},
+				},
+			},
 		}
 		err := rules.Validate(l, tax.AddonContext(en16931.V2017))
 		assert.NoError(t, err)
@@ -263,6 +272,15 @@ func TestValidateBillDiscount(t *testing.T) {
 				untdid.ExtKeyAllowance: "67",
 			},
 			Amount: num.MakeAmount(100, 2),
+			Taxes: tax.Set{
+				{
+					Category: tax.CategoryVAT,
+					Rate:     "standard",
+					Ext: tax.Extensions{
+						untdid.ExtKeyTaxCategory: en16931.TaxCategoryStandard,
+					},
+				},
+			},
 		}
 		err := rules.Validate(l, tax.AddonContext(en16931.V2017))
 		assert.NoError(t, err)
@@ -271,6 +289,15 @@ func TestValidateBillDiscount(t *testing.T) {
 	t.Run("without reason or extension", func(t *testing.T) {
 		l := &bill.Discount{
 			Amount: num.MakeAmount(100, 2),
+			Taxes: tax.Set{
+				{
+					Category: tax.CategoryVAT,
+					Rate:     "standard",
+					Ext: tax.Extensions{
+						untdid.ExtKeyTaxCategory: en16931.TaxCategoryStandard,
+					},
+				},
+			},
 		}
 		err := rules.Validate(l, tax.AddonContext(en16931.V2017))
 		assert.ErrorContains(t, err, "either a reason or an allowance type extension is required")
@@ -283,9 +310,27 @@ func TestValidateBillDiscount(t *testing.T) {
 				untdid.ExtKeyAllowance: "67",
 			},
 			Amount: num.MakeAmount(100, 2),
+			Taxes: tax.Set{
+				{
+					Category: tax.CategoryVAT,
+					Rate:     "standard",
+					Ext: tax.Extensions{
+						untdid.ExtKeyTaxCategory: en16931.TaxCategoryStandard,
+					},
+				},
+			},
 		}
 		err := rules.Validate(l, tax.AddonContext(en16931.V2017))
 		assert.NoError(t, err)
+	})
+
+	t.Run("without taxes (BR-32)", func(t *testing.T) {
+		l := &bill.Discount{
+			Reason: "Product sample",
+			Amount: num.MakeAmount(100, 2),
+		}
+		err := rules.Validate(l, tax.AddonContext(en16931.V2017))
+		assert.ErrorContains(t, err, "taxes are required (BR-32)")
 	})
 }